migrate test_keystone to pytest
